refactor: 부가적인 utils 목적의 클래스들은 utils 패키지 밑으로 옮김
